#-: Experiments with drawlist and dearpygui

+++ src/main.py
@@ -4,35 +4,58 @@ import time
 
 def main():
     dpg.create_context()
+
+    with dpg.window(label="Snake Game", width=50, height=50, tag="main_window"):
+        with dpg.drawlist(width=-1, height=-1, tag="drawlist"):
+            dpg.draw_rectangle(pmin=(0, 0), pmax=(100, 100), color=(255, 0, 0, 255))
+
     dpg.create_viewport()
     dpg.setup_dearpygui()
     dpg.show_viewport()
-
-    dpg.add_window(label="tutorial", width=500, height=500, tag="main_window")
     dpg.set_primary_window(window="main_window", value=True)
     
+    squares = []
     while dpg.is_dearpygui_running():
-        drawlist_size = (dpg.get_item_width("main_window") - 20, dpg.get_item_height("main_window") - 20)
-
-        dpg.delete_item("drawlist")
-        dpg.add_drawlist(width=drawlist_size[0], height=drawlist_size[1], tag="drawlist", parent="main_window")
-        
-        GRID_SIZE = 10
-        SQUARE_SIZE = min(drawlist_size[0] // GRID_SIZE, drawlist_size[1] // GRID_SIZE)
-        
-        squares = [[((SQUARE_SIZE * i, SQUARE_SIZE * j), (SQUARE_SIZE + SQUARE_SIZE * i, SQUARE_SIZE + SQUARE_SIZE * j)) for i in range(GRID_SIZE)] for j in range(GRID_SIZE)]
-        for i in range(GRID_SIZE):
-            for j in range(GRID_SIZE):
-                dpg.delete_item(f"square_{i}_{j}")
-                dpg.draw_rectangle(squares[i][j][0], squares[i][j][1], tag=f"square_{i}_{j}", parent="drawlist", fill=(255, 0, 0))
-        
         fps = 60
         frame_time = 1 / fps
         time.sleep(frame_time)
-            
         dpg.render_dearpygui_frame()
         
+        window_width = dpg.get_item_width("main_window")
+        window_height = dpg.get_item_height("main_window")
+        dpg.configure_item("drawlist", width=window_width, height=window_height)
+        
+        if window_width is None or window_height is None:
+            continue
+        
+        SQUARE_SIZE = 50
+        GRID_SIZE = min(window_width // SQUARE_SIZE, window_height // SQUARE_SIZE)
+
+        for i in range(len(squares)):
+            for j in range(len(squares[0])):
+                dpg.delete_item(f"square_{i}_{j}")
+        squares = [
+            [
+                (
+                    (SQUARE_SIZE * i, SQUARE_SIZE * j),
+                    (SQUARE_SIZE + SQUARE_SIZE * i, SQUARE_SIZE + SQUARE_SIZE * j),
+                )
+                for i in range(GRID_SIZE)
+            ]
+            for j in range(GRID_SIZE)
+        ]
+        for i in range(len(squares)):
+            for j in range(len(squares[0])):
+                dpg.draw_rectangle(
+                    squares[i][j][0],
+                    squares[i][j][1],
+                    tag=f"square_{i}_{j}",
+                    parent="drawlist",
+                    fill=(255, 0, 0),
+                )
+
     dpg.destroy_context()
+
 
 if __name__ == "__main__":
     main()
